fix(usepdfupload): refactor into ordinary functions
